+++ build-conventions-secant/src/main/kotlin/model/ZashiFlavorType.kt
@@ -1,0 +1,15 @@
+package model
+
+const val ZASHI_FLAVOR_DIMENSION = "network"
+
+// We would rather name them "testnet" and "mainnet" but product flavor names cannot start with the word "test"
+sealed class ZashiFlavorType {
+    abstract val name: String
+
+    object Mainnet : ZashiBuildType() {
+        override val name = "zcashmainnet"
+    }
+    object Testnet : ZashiBuildType() {
+        override val name = "zcashtestnet"
+    }
+}